// src/api/models/RouteGeometry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Distrologiq
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RouteGeometry
 */
export interface RouteGeometry {
    /**
     * 
     * @type {string}
     * @memberof RouteGeometry
     */
    type: string;
    /**
     * 
     * @type {string}
     * @memberof RouteGeometry
     */
    coordinates: string;
}

export function RouteGeometryFromJSON(json: any): RouteGeometry {
    return RouteGeometryFromJSONTyped(json, false);
}

export function RouteGeometryFromJSONTyped(json: any, ignoreDiscriminator: boolean): RouteGeometry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'type': json['type'],
        'coordinates': json['coordinates'],
    };
}

export function RouteGeometryToJSON(value?: RouteGeometry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'type': value.type,
        'coordinates': value.coordinates,
    };
}