// src/api/models/Settings.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Distrologiq
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Destination,
    DestinationFromJSON,
    DestinationFromJSONTyped,
    DestinationToJSON,
} from './';

/**
 * 
 * @export
 * @interface Settings
 */
export interface Settings {
    /**
     * 
     * @type {string}
     * @memberof Settings
     */
    uuid: string;
    /**
     * 
     * @type {string}
     * @memberof Settings
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof Settings
     */
    avgLoadTime: number;
    /**
     * 
     * @type {Destination}
     * @memberof Settings
     */
    destination: Destination;
    /**
     * 
     * @type {Date}
     * @memberof Settings
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Settings
     */
    updatedAt: Date;
}

export function SettingsFromJSON(json: any): Settings {
    return SettingsFromJSONTyped(json, false);
}

export function SettingsFromJSONTyped(json: any, ignoreDiscriminator: boolean): Settings {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'uuid': json['uuid'],
        'name': json['name'],
        'avgLoadTime': json['avgLoadTime'],
        'destination': DestinationFromJSON(json['destination']),
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function SettingsToJSON(value?: Settings | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'uuid': value.uuid,
        'name': value.name,
        'avgLoadTime': value.avgLoadTime,
        'destination': DestinationToJSON(value.destination),
        'createdAt': (value.createdAt.toISOString()),
        'updatedAt': (value.updatedAt.toISOString()),
    };
}